Use 10 symbols for all stream tests
