+++ src/streams/coinbase/mod.rs
@@ -92,46 +92,3 @@ impl CoinbaseClientBuilder {
     }
 }
 
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use futures::StreamExt;
-    use tokio::time::{Duration, timeout};
-
-    #[tokio::test]
-    #[ignore]
-    async fn test_coinbase_stream_event() {
-        let symbols = vec![
-            "BTC-USD",
-            "ETH-USD",
-            "SOL-USD",
-            "ADA-USD",
-            "DOGE-USD",
-            "AVAX-USD",
-            "XRP-USD",
-            "LTC-USD",
-            "LINK-USD",
-            "MATIC-USD",
-        ];
-        let client = CoinbaseClient::builder()
-            .add_symbols(symbols)
-            .build()
-            .unwrap();
-        let mut stream = client.stream_events().await.unwrap();
-        let count = timeout(Duration::from_secs(30), async {
-            let mut received = 0;
-            while let Some(event) = stream.next().await {
-                if event.is_ok() {
-                    received += 1;
-                    if received >= 50 {
-                        break;
-                    }
-                }
-            }
-            received
-        })
-        .await
-        .expect("timed out waiting for events");
-        assert!(count >= 50, "insufficient events: {}", count);
-    }
-}
